Feature #13197: implementing new community component.
Refactor some duplication of codes in Silverpeas component definition
(feedback from SonarQube analysis)
